+++ changelog.md
@@ -13,6 +13,8 @@ Codion Change Log
 - SLF4J dependency added, DefaultDatabaseConnection logging improved.
 - ConnectionPoolWrapper.DEFAULT_CHECK_OUT_TIMEOUT added.
 - ConnectionPoolWrapper.VALIDATE_CONNECTIONS_ON_CHECKOUT added.
+### is.codion.common.model
+- DefaultFilterModelItems.filterIncremental() added.
 ### is.codion.dbms.h2
 - H2Database, security validation for the H2 database script injection.
 ### is.codion.swing.common.ui